refactor: minor cleanups, style fixes
