+++ app/src/main/java/com/example/cupcake/MainComposeActivity.kt
@@ -3,13 +3,13 @@ package com.example.cupcake
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
-import androidx.activity.enableEdgeToEdge
 import androidx.activity.viewModels
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.padding
 import androidx.compose.material3.Scaffold
 import androidx.compose.ui.Modifier
 import com.example.cupcake.model.OrderViewModel
+import com.example.cupcake.navigation.Navigation
 import com.example.cupcake.ui.theme.CupcakeTheme
 
 class MainComposeActivity : ComponentActivity() {
