+++ app/src/main/java/com/example/cupcake/ui/screens/PickupScreen.kt
@@ -54,7 +54,7 @@ fun PickupScreen(
 }
 
 @Composable
-fun PickupScreenContent(
+private fun PickupScreenContent(
     sharedViewModel: OrderViewModel,
     onNavigateNext: () -> Unit,
     onNavigateToStart: () -> Unit,
